Compressed fpre2

+++ cosmictiger/sph_cuda.hpp
@@ -70,7 +70,7 @@ struct sph_run_cuda_data {
 	float* vy;
 	float* vz;
 	float16* fpre1;
-	float* fpre2;
+	float16* fpre2;
 	float* pre;
 	float* h;
 	char* rungs;
